Create Scaled Band Unfolding
This code is an example of a scaled band unfolding on HfSe2 material.
If anyone here has done a band unfolding, it is common for the band structure of the supercell to be off scale to the base cell hence an offset is created. This code aims to fix this.